Sync esp-tflite-micro from github - 592464

See merge request app-frameworks/esp-tflite-micro!123

+++ examples/micro_speech/main/command_responder.cc
@@ -20,8 +20,8 @@ limitations under the License.
 // to the error console. Real applications will want to take some custom
 // action instead, and should implement their own versions of this function.
 void RespondToCommand(int32_t current_time, const char* found_command,
-                      uint8_t score, bool is_new_command) {
+                      float score, bool is_new_command) {
   if (is_new_command) {
-    MicroPrintf("Heard %s (%d) @%dms", found_command, score, current_time);
+    MicroPrintf("Heard %s (%.4f) @%dms", found_command, score, current_time);
   }
 }
