+++ examples/micro_speech/main/command_responder.h
@@ -25,6 +25,6 @@ limitations under the License.
 // argument, `score` has the numerical confidence, and `is_new_command` is set
 // if the previous command was different to this one.
 void RespondToCommand(int32_t current_time, const char* found_command,
-                      uint8_t score, bool is_new_command);
+                      float score, bool is_new_command);
 
 #endif  // TENSORFLOW_LITE_MICRO_EXAMPLES_MICRO_SPEECH_COMMAND_RESPONDER_H_
